+++ R/generics.R
@@ -9,3 +9,7 @@ broom::glance
 #' @importFrom broom augment
 #' @export
 broom::augment
+
+#' @importFrom lme4 VarCorr
+#' @export
+lme4::VarCorr